refactor(harness): centralize session lifecycle operations
